Don't send template parameters with empty values
Parameters which are not set by the user in the service portal
dialog are being sent to us as empty strings. We should not be sending
those parameters over to OpenShift. This is the way the OpenShift
catalog UI handles the case when a user doesn't fill in a non-required
parameter.